# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Takeaway App API",
        "description": "API documentation for the Takeaway App",
        "contact": {
            "email": "linwenhao5@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders for the authenticated customer (paginated)",
                "description": "Get all orders for the authenticated customer (with pagination).",
                "operationId": "7b1be3658ec01b201e7c9106f4b57cd1",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of orders per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orders retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "orders": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order and get payment URL",
                "description": "Create a new order and return payment URL.",
                "operationId": "ba810fa4904946d560e892ea4c047502",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_type",
                                    "reserve_time"
                                ],
                                "properties": {
                                    "order_type": {
                                        "description": "Order type: delivery or pickup",
                                        "type": "string",
                                        "enum": [
                                            "delivery",
                                            "pickup"
                                        ]
                                    },
                                    "address_id": {
                                        "description": "Required if order_type is delivery. Address ID for delivery orders",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "reserve_time": {
                                        "description": "Reserve time for delivery or pickup",
                                        "type": "string",
                                        "example": "2025-09-06 18:30"
                                    },
                                    "platform": {
                                        "description": "Platform (optional)",
                                        "type": "string",
                                        "example": "web"
                                    },
                                    "host": {
                                        "description": "Request host (optional)",
                                        "type": "string",
                                        "example": "example.com"
                                    },
                                    "note": {
                                        "description": "Additional note for the order (optional)",
                                        "type": "string",
                                        "example": "Please add extra wasabi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "order_id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "payment_url": {
                                            "type": "string",
                                            "example": "https://www.mollie.com/paymentscreen/example"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{orderId}/repay": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Repay an unpaid order and get a new payment URL",
                "description": "Repay an unpaid order and get a new payment URL.",
                "operationId": "67a6206b18ff1004d647862b15994917",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of the order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment URL generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "order_id": {
                                            "type": "integer",
                                            "example": 123
                                        },
                                        "payment_url": {
                                            "type": "string",
                                            "example": "https://www.mollie.com/paymentscreen/example"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Order already paid or not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token:",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}